05/05.test.ts
import * as path from "@std/path";
import { expect } from "@std/expect";
import { describe, it } from "@std/testing/bdd";
import { solveA, solveB } from "./05.ts";

const currentDay = path.basename(Deno.cwd());

describe(`AOC 2024 - Day ${currentDay}`, () => {
	describe("Part A", () => {
		it("Example", () => {
			expect(solveA("example_a", currentDay)).toBe(143);
		});

		it("Solution", () => {
			expect(solveA("input", currentDay)).toBe(4569);
		});
	});

	describe("Part B", () => {
		it("Example", () => {
			expect(solveB("example_b", currentDay)).toBe(123);
		});

		it("Solution", () => {
			expect(solveB("input", currentDay)).toBe(6456);
		});
	});
});
